Module cleanup.
Moved scraping utility functions into a separate file (this also breaks
a circular dependency with fetch_events, local.*.scrapers and config)

Removed backend.* from modules so code can be run directly from backend

Replaced hardcoded value in index_elasticsearch.py
