+++ backend/config.py
@@ -3,10 +3,12 @@ import os
 # Module containing scrapers for extracting data from webpages
 # Every public function (i.e., those not beginning with _) should return a list
 # of event dictionaries when called with no arguments.
-import backend.local.upenn.scrapers as scrapers
+import local.upenn.scrapers
+scrapers = local.upenn.scrapers
 
 # Module containing a list of locations (for feature generation)
-from backend.local.upenn.locations import locations
+import local.upenn.locations
+locations = local.upenn.locations.locations
 
 # Path to save event data (as json)
 PATH_DATA = os.path.join('data', 'events.json')  # path to save events
